Account.class added to store account info, the AccountController takes the object Account to use throughout program
