+++ .idea/misc.xml
@@ -4,8 +4,8 @@
     <option name="filePathToZoomLevelMap">
       <map>
         <entry key="app/src/main/res/drawable/ic_launcher_background.xml" value="0.14814814814814814" />
-        <entry key="app/src/main/res/layout/activity_create_account.xml" value="0.165" />
-        <entry key="app/src/main/res/layout/activity_homepage.xml" value="0.153125" />
+        <entry key="app/src/main/res/layout/activity_create_account.xml" value="0.25" />
+        <entry key="app/src/main/res/layout/activity_homepage.xml" value="0.25" />
         <entry key="app/src/main/res/layout/activity_main.xml" value="0.25" />
         <entry key="app/src/main/res/layout/create_account_layout.xml" value="0.15989583333333332" />
       </map>
