+++ app/src/main/java/com/example/chefskiss2/Account.java
@@ -1,0 +1,26 @@
+package com.example.chefskiss2;
+
+
+public class Account {
+    private String username;
+    private String email;
+    private String password;
+
+    public Account(String username, String email, String password) {
+        this.username = username;
+        this.email = email;
+        this.password = password;
+    }
+
+    public String getEmail() {
+        return this.email;
+    }
+
+    public String getUsername() {
+        return this.username;
+    }
+
+    public String getPassword() {
+        return this.password;
+    }
+}
