+++ app/src/main/java/com/example/chefskiss2/CreateAccount.java
@@ -25,20 +25,20 @@ public class CreateAccount extends AppCompatActivity {
         EditText password = (EditText) findViewById(R.id.password);
         TextView errorMessage = (TextView) findViewById(R.id.invalidInfoMessage);
 
+        String emailString = email.getText().toString();
+        String usernameString = username.getText().toString();
+        String passwordString = password.getText().toString();
+
         Button createAccount = (Button) findViewById(R.id.createAccountBtn2);
-        AccountController AC = new AccountController();
+
 
         createAccount.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                String emailString = email.getText().toString();
-                String usernameString = username.getText().toString();
-                String passwordString = password.getText().toString();
+
 
                 //*** Add later if username is already taken once DB is implemented
                 if (emailString.contains("@") && emailString.endsWith(".com")) {
-                    AC.createAccount(emailString, usernameString, passwordString);
-                    CreateAccount.this.AC = AC;
 
                     //Used to check a password contains uppercase, lowercase,
                     // has a special character, and a number
@@ -55,6 +55,7 @@ public class CreateAccount extends AppCompatActivity {
 
                     //Checks if the password matches
                     if(m.matches()) {
+                        CreateAccount.this.AC = new AccountController(new Account(usernameString, emailString, passwordString));
 
                         //If account is stored in AC
                         if (CreateAccount.this.AC.isLoggedIn()) {
