+++ app/src/main/java/com/example/chefskiss2/Homepage.java
@@ -3,12 +3,27 @@ package com.example.chefskiss2;
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.os.Bundle;
+import android.view.View;
+import android.widget.ImageButton;
 
 public class Homepage extends AppCompatActivity {
+
+    public Homepage() {
+
+    }
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_homepage);
+
+        ImageButton recipesButton = (ImageButton) findViewById(R.id.imageButton3);
+
+        recipesButton.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+
+            }
+        });
     }
 }
